Adding classes AABB and Scene
