+++ frontend/web/src/assets/lib/jsc3d/jsc3d.js
@@ -33,130 +33,6 @@ var JSC3D = JSC3D || {};
 
 
 
-/**
-	@class Scene
-
-	This class implements scene that contains a group of meshes that forms the world.
- */
-JSC3D.Scene = function(name) {
-	this.name = name || '';
-	this.srcUrl = '';
-	this.aabb = null;
-	this.children = [];
-	this.maxChildId = 1;
-};
-
-/**
-	Initialize the scene.
- */
-JSC3D.Scene.prototype.init = function() {
-	if(this.isEmpty())
-		return;
-
-	for(var i=0; i<this.children.length; i++)
-		this.children[i].init();
-
-	if(!this.aabb) {
-		this.aabb = new JSC3D.AABB;
-		this.calcAABB();
-	}
-};
-
-/**
-	See if the scene is empty.
-	@returns {Boolean} true if it does not contain meshes; false if it has any.
- */
-JSC3D.Scene.prototype.isEmpty = function() {
-	return (this.children.length == 0);
-};
-
-/**
-	Add a mesh to the scene.
-	@param {JSC3D.Mesh} mesh the mesh to be added.
- */
-JSC3D.Scene.prototype.addChild = function(mesh) {
-	mesh.internalId = this.maxChildId++;
-	this.children.push(mesh);
-};
-
-/**
-	Remove a mesh from the scene.
-	@param {JSC3D.Mesh} mesh the mesh to be removed.
- */
-JSC3D.Scene.prototype.removeChild = function(mesh) {
-	var foundAt = this.children.indexOf(mesh);
-	if(foundAt >= 0)
-		this.children.splice(foundAt, 1);
-};
-
-/**
-	Get all meshes in the scene.
-	@returns {Array} meshes as an array.
- */
-JSC3D.Scene.prototype.getChildren = function() {
-	return this.children.slice(0);
-};
-
-/**
-	Traverse meshes in the scene, calling a given function on each of them.
-	@param {Function} operator a function that will be called on each mesh.
- */
-JSC3D.Scene.prototype.forEachChild = function(operator) {
-	if((typeof operator) != 'function')
-		return;
-
-	for(var i=0; i<this.children.length; i++) {
-		if(operator.call(null, this.children[i]))
-			break;
-	}
-};
-
-/**
-	Calculate AABB of the scene.
-	@private
- */
-JSC3D.Scene.prototype.calcAABB = function() {
-	this.aabb.minX = this.aabb.minY = this.aabb.minZ = Infinity;
-	this.aabb.maxX = this.aabb.maxY = this.aabb.maxZ = -Infinity;
-	for(var i=0; i<this.children.length; i++) {
-		var child = this.children[i];
-		if(!child.isTrivial()) {
-			var minX = child.aabb.minX;
-			var minY = child.aabb.minY;
-			var minZ = child.aabb.minZ;
-			var maxX = child.aabb.maxX;
-			var maxY = child.aabb.maxY;
-			var maxZ = child.aabb.maxZ;
-			if(this.aabb.minX > minX)
-				this.aabb.minX = minX;
-			if(this.aabb.minY > minY)
-				this.aabb.minY = minY;
-			if(this.aabb.minZ > minZ)
-				this.aabb.minZ = minZ;
-			if(this.aabb.maxX < maxX)
-				this.aabb.maxX = maxX;
-			if(this.aabb.maxY < maxY)
-				this.aabb.maxY = maxY;
-			if(this.aabb.maxZ < maxZ)
-				this.aabb.maxZ = maxZ;
-		}
-	}
-};
-
-/**
- * {String} Name of the scene.
- */
-JSC3D.Scene.prototype.name = '';
-/**
- * {String} Source URL of the scene, empty if none. Read only.
- */
-JSC3D.Scene.prototype.srcUrl = '';
-/**
- * {JSC3D.AABB} The Axis-aligned bounding box of the whole scene. Read only.
- */
-JSC3D.Scene.prototype.aabb = null;
-JSC3D.Scene.prototype.children = null;
-JSC3D.Scene.prototype.maxChildId = 1;
 
 
 /**
@@ -600,66 +476,6 @@ JSC3D.Mesh.prototype.transformedVertexNormalBuffer = null;
 
 
 
-
-
-/**
-	@class AABB
-
-	This class implements the Axis-Aligned Bounding Box to measure spatial enclosure.
- */
-JSC3D.AABB = function() {
-	/**
-	 * {Number} X coordinate of the minimum edge of the box.
-	 */
-	this.minX = 0;
-	/**
-	 * {Number} Y coordinate of the minimum edge of the box.
-	 */
-	this.minY = 0;
-	/**
-	 * {Number} Z coordinate of the minimum edge of the box.
-	 */
-	this.minZ = 0;
-	/**
-	 * {Number} X coordinate of the maximum edge of the box.
-	 */
-	this.maxX = 0;
-	/**
-	 * {Number} Y coordinate of the maximum edge of the box.
-	 */
-	this.maxY = 0;
-	/**
-	 * {Number} Z coordinate of the maximum edge of the box.
-	 */
-	this.maxZ = 0;
-};
-
-/**
-	Get center coordinates of the AABB.
-	@param {Array} c an array to receive the result.
-	@returns {Array} center coordinates as an array.
- */
-JSC3D.AABB.prototype.center = function(c) {
-	if(c) {
-		c[0] = 0.5 * (this.minX + this.maxX);
-		c[1] = 0.5 * (this.minY + this.maxY);
-		c[2] = 0.5 * (this.minZ + this.maxZ);
-	}
-	else
-		c = [0.5 * (this.minX + this.maxX), 0.5 * (this.minY + this.maxY), 0.5 * (this.minZ + this.maxZ)];
-	return c;
-};
-
-/**
-	Get the length of the diagonal of the AABB.
-	@returns {Number} length of the diagonal.
- */
-JSC3D.AABB.prototype.lengthOfDiagonal = function() {
-	var xx = this.maxX - this.minX;
-	var yy = this.maxY - this.minY;
-	var zz = this.maxZ - this.minZ;
-	return Math.sqrt(xx * xx + yy * yy + zz * zz);
-};
